requirements.txt
Flask~=3.0
Flask-SQLAlchemy~=3.0
Flask-Migrate~=4.0
psycopg2-binary~=2.9
python-dotenv~=1.0
pytest~=8.0
